Prevent empty Kirki::$fields
This is a fix/suggestion related to: https://github.com/kirki-framework/field-typography/issues/2

Currently the global variable `Kirki::$fields` stays empty as single control packages (e.g. `kirki-framework/control-select`) do not add the `$args` to it. 

As the usage of `Kirki/Compatibility` package seems to be optional, I guess this is a good position to keep it seperated, valid for all fields as it adds a field to the global variable if a separated class exists for the field/control type. 

Best regards!